+++ test/unit/auth/auth-api-request.spec.ts
@@ -36,7 +36,8 @@ import {
   RESERVED_CLAIMS, FIREBASE_AUTH_UPLOAD_ACCOUNT, FIREBASE_AUTH_CREATE_SESSION_COOKIE,
   EMAIL_ACTION_REQUEST_TYPES, TenantAwareAuthRequestHandler, AbstractAuthRequestHandler,
 } from '../../../src/auth/auth-api-request';
-import { UserImportBuilder, UserImportRecord } from '../../../src/auth/user-import-builder';
+import { UserImportRecord } from '../../../src/auth/user-import-builder';
+import { UserImportBuilder } from '../../../src/auth/user-import-builder-internal';
 import { AuthClientErrorCode, FirebaseAuthError } from '../../../src/utils/error';
 import { ActionCodeSettingsBuilder } from '../../../src/auth/action-code-settings-builder';
 import {
